Finish resource manager assignment.

+++ Source/Engine/Renderer/Renderer.cpp
@@ -75,8 +75,8 @@ namespace ane {
 		SDL_Rect dest;
 		dest.x = (int)(x - (size.x * 0.5f));
 		dest.y = (int)(y - (size.y * 0.5f));
-		dest.w = size.x;
-		dest.h = size.y;
+		dest.w = (int)size.x;
+		dest.h = (int)size.y;
 
 		// https://wiki.libsdl.org/SDL2/SDL_RenderCopyEx
 		SDL_RenderCopyEx(this->renderer, texture->texture, NULL, &dest, angle, NULL, SDL_FLIP_NONE);
